[Huawei][AscendNPU IR]fix cleancode warning

Created-by: tianjihua
Commit-by: tianjihua
Merged-by: zyf25
Description: [Huawei][AscendNPU IR]fix cleancode warning

See merge request: Ascend/AscendNPU-IR-Dev!522

+++ bishengir/include/bishengir/Tools/bishengir-compile/Config.h
@@ -423,7 +423,7 @@ public:
   unsigned blockDim() const { return blockDimFlag; }
 
   BiShengIRCompileMainConfig &setVfMergeLevel(unsigned enableVfMergeLevel) {
-    enableVfMergeLevelFlag = enableVfMergeLevel;
+    enableVfMergeLevelFlag = (int32_t)enableVfMergeLevel;
     return *this;
   }
   int32_t enableVfMergeLevel() const { return enableVfMergeLevelFlag; }
